requirements.txt
python==3.8.5
pytorch==1.6.0
numpy==1.19.1